middleware: Handler interface embeds new Nexter interface
Nexter allows middleware chained together to be treated like a singly-linked list, since the "Next" handler in the chain can be read and written to. This allows the middleware stack to change after it is compiled. These getter/setter functions are one-liners, and having middleware.Handler defined with these two methods gives us compile-time safety rather than relying on type assertions and "remembering" to add those tiny methods to every middleware. It may be inconvenient for external users of the middleware chain who write their own middleware, but this is in the best interest of Caddy and all others who want to make easily mutable middleware chains.