Swappable server config with API; bad config causes rollback
If the user PUTs a new Caddyfile to use, replacing all existing configuration, then the listeners are gracefully closed and the new listeners will take their place. If they fail to start correctly, the previous configuration is rolled back into place. The rollback may still fail due to system environment changes, but this failure is more unlikely since it was guaranteed to be working before.